Added Mouse Collision Detection Triggers
Added the ability to trigger events based on collision with a bounding box.  The UICTrigger component will add a bounding box that checks for collision with the mouse, and you can add 3 events to activate those, including on mouse just entered, on mouse still inside the box, and on mouse just exited. I do not see at the moment any reason to add a still outside box, so I did not add that one the event on just entered will only fire when it has just entered the box. This is good for setup functions or anything else that is needing to be run as soon as it enters, like maybe the start of a timer for detecting how long the box has been in there this event will only run once and will reset only when the mouse exits and enters again. the still inside event on the other hand will run every frame while the box is still in the box this is where we would do all the main code. I showed in the test class how we can detect clicking in the box very easily with this function. The last event is on exit and this will fire only one frame as soon as the mouse exits the bounding box. I also started adding the ability to add scenes to just test how  I can disable and enable UI entites in the scene. Every UI element will be it's own entity and borrows form the regular enitiy functionality so that I can just add components as needed for functionality.